moved functions related to Context.context to a proper directory
