+++ cxts/RecieveFromWorkTimeout.go
@@ -1,4 +1,4 @@
-package goroutines
+package cxts
 
 import (
 	"context"
